Initial version
The repo starts with three new TestContainer containers:

- SAP HANA
- Vertica
- SQLite

The first two are essentially cloud databases that provide docker
images for testing and development.

SQLite is an embedded database that relies on host resources - a docker
container is pointless. However it's a widely used database and creating
a fake container allows developers to use tools developed for the larger
TestContainer ecosystem.