Added tc_builder
Added tc_builder - a module that can build the 'testcontainer' software
within a docker file.

Note: it is incomplete - it only contains the Dockerfile and the files
required to download gradle 8.x. It can run 'gradle' but not 'gradle build'.
The java code, e.g., for the Container, will be added later.